+++ slither/detectors/abstract_detector.py
@@ -7,8 +7,12 @@ from slither.core.compilation_unit import SlitherCompilationUnit, Language
 from slither.core.declarations import Contract
 from slither.formatters.exceptions import FormatImpossible
 from slither.formatters.utils.patches import apply_patch, create_diff
-from slither.utils.colors import green, yellow, red
-from slither.utils.comparable_enum import ComparableEnum
+from slither.utils.colors import yellow
+from slither.detectors.classification import (
+    classification_txt,
+    classification_colors,
+    DetectorClassification,
+)
 from slither.utils.output import Output
 
 if TYPE_CHECKING:
@@ -18,33 +22,6 @@ if TYPE_CHECKING:
 
 class IncorrectDetectorInitialization(Exception):
     pass
-
-
-class DetectorClassification(ComparableEnum):
-    HIGH = 0
-    MEDIUM = 1
-    LOW = 2
-    INFORMATIONAL = 3
-    OPTIMIZATION = 4
-
-    UNIMPLEMENTED = 999
-
-
-classification_colors: Dict[DetectorClassification, Callable[[str], str]] = {
-    DetectorClassification.INFORMATIONAL: green,
-    DetectorClassification.OPTIMIZATION: green,
-    DetectorClassification.LOW: green,
-    DetectorClassification.MEDIUM: yellow,
-    DetectorClassification.HIGH: red,
-}
-
-classification_txt = {
-    DetectorClassification.INFORMATIONAL: "Informational",
-    DetectorClassification.OPTIMIZATION: "Optimization",
-    DetectorClassification.LOW: "Low",
-    DetectorClassification.MEDIUM: "Medium",
-    DetectorClassification.HIGH: "High",
-}
 
 
 def make_solc_versions(minor: int, patch_min: int, patch_max: int) -> List[str]:
